+++ ton_client/src/net/mod.rs
@@ -14,30 +14,32 @@
 use crate::client::ClientContext;
 use crate::error::ClientResult;
 
+pub(crate) mod aggregates;
 mod errors;
 mod gql;
 pub(crate) mod queries;
-mod websocket_link;
-mod server_link;
 mod server_info;
+mod server_link;
 pub(crate) mod subscriptions;
 mod types;
+mod websocket_link;
 
+pub use aggregates::{
+    aggregate_collection, ParamsOfAggregateCollection, ResultOfAggregateCollection,
+};
 pub use errors::{Error, ErrorCode};
-pub use gql::{OrderBy, SortDirection};
+pub use gql::{AggregationFn, FieldAggregation, OrderBy, SortDirection};
 pub use queries::{
-    query, query_collection, wait_for_collection,
-    ParamsOfQuery, ParamsOfQueryCollection, ParamsOfWaitForCollection, ResultOfQuery,
-    ResultOfQueryCollection, ResultOfWaitForCollection,
+    query, query_collection, wait_for_collection, ParamsOfQuery, ParamsOfQueryCollection,
+    ParamsOfWaitForCollection, ResultOfQuery, ResultOfQueryCollection, ResultOfWaitForCollection,
 };
 pub use subscriptions::{
-    subscribe_collection, unsubscribe,
-    ParamsOfSubscribeCollection, ResultOfSubscribeCollection, ResultOfSubscription,
-    SubscriptionResponseType
+    subscribe_collection, unsubscribe, ParamsOfSubscribeCollection, ResultOfSubscribeCollection,
+    ResultOfSubscription, SubscriptionResponseType,
 };
 pub use types::{
-    NetworkConfig,
-    MESSAGES_TABLE_NAME, CONTRACTS_TABLE_NAME, BLOCKS_TABLE_NAME, TRANSACTIONS_TABLE_NAME
+    NetworkConfig, BLOCKS_TABLE_NAME, CONTRACTS_TABLE_NAME, MESSAGES_TABLE_NAME,
+    TRANSACTIONS_TABLE_NAME,
 };
 
 pub(crate) use server_link::{ServerLink, MAX_TIMEOUT};
@@ -47,18 +49,14 @@ mod tests;
 
 /// Suspends network module to stop any network activity
 #[api_function]
-pub async fn suspend(
-    context: std::sync::Arc<ClientContext>,
-) -> ClientResult<()> {
+pub async fn suspend(context: std::sync::Arc<ClientContext>) -> ClientResult<()> {
     context.get_server_link()?.suspend().await;
     Ok(())
 }
 
 /// Resumes network module to enable network activity
 #[api_function]
-pub async fn resume(
-    context: std::sync::Arc<ClientContext>,
-) -> ClientResult<()> {
+pub async fn resume(context: std::sync::Arc<ClientContext>) -> ClientResult<()> {
     context.get_server_link()?.resume().await;
     Ok(())
 }
@@ -83,11 +81,11 @@ pub async fn find_last_shard_block(
 ) -> ClientResult<ResultOfFindLastShardBlock> {
     let address = crate::encoding::account_decode(&params.address)?;
 
-    let block_id = crate::processing::blocks_walking::find_last_shard_block(&context, &address)
-        .await?;
+    let block_id =
+        crate::processing::blocks_walking::find_last_shard_block(&context, &address).await?;
 
     Ok(ResultOfFindLastShardBlock {
-        block_id: block_id.to_string()
+        block_id: block_id.to_string(),
     })
 }
 
@@ -99,13 +97,11 @@ pub struct EndpointsSet {
 
 /// Requests the list of alternative endpoints from server
 #[api_function]
-pub async fn fetch_endpoints(
-    context: std::sync::Arc<ClientContext>,
-) -> ClientResult<EndpointsSet> {
+pub async fn fetch_endpoints(context: std::sync::Arc<ClientContext>) -> ClientResult<EndpointsSet> {
     let client = context.get_server_link()?;
-    
+
     Ok(EndpointsSet {
-        endpoints: client.fetch_endpoints().await?
+        endpoints: client.fetch_endpoints().await?,
     })
 }
 
@@ -113,16 +109,16 @@ pub async fn fetch_endpoints(
 #[api_function]
 pub async fn set_endpoints(
     context: std::sync::Arc<ClientContext>,
-    params: EndpointsSet
+    params: EndpointsSet,
 ) -> ClientResult<()> {
     if params.endpoints.len() == 0 {
         return Err(Error::no_endpoints_provided());
     }
-    
+
     context
         .get_server_link()?
         .set_endpoints(params.endpoints)
         .await;
-    
+
     Ok(())
 }