+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Release Notes
 All notable changes to this project will be documented in this file.
 
+## 1.6.0 Jan 20, 2020
+### New
+- `nacl_sign_detached_verify` function to verify detached signature.
+- `aggregate_collection` function as a wrapper for GraphQL aggregation queries.
+
+### Fixed
+- device time synchronization is checked only in `send_message`. Data quering does not require proper time
 
 ## 1.5.2 Dec 30, 2020